CVS_SILENT Minor fix. It's the first case I see an overflow in 'float' that needs to be fixed using 'double' type.
svn path=/trunk/kdeextragear-1/amarok/; revision=289810

+++ amarok/analyzers/glanalyzer2.cpp
@@ -91,7 +91,7 @@ void GLAnalyzer2::resizeGL( int w, int h )
     // Get current timestamp.
     timeval tv;
     gettimeofday( &tv, NULL );
-    show.timeStamp = (float)tv.tv_sec + (float)tv.tv_usec/1000000.0;
+    show.timeStamp = (double)tv.tv_sec + (double)tv.tv_usec/1000000.0;
 }
 
 void GLAnalyzer2::paused()
@@ -145,7 +145,7 @@ void GLAnalyzer2::paintGL()
     // Compute the dT since the last call to paintGL and update timings
     timeval tv;
     gettimeofday( &tv, NULL );
-    double currentTime = (float)tv.tv_sec + (float)tv.tv_usec/1000000.0;
+    double currentTime = (double)tv.tv_sec + (double)tv.tv_usec/1000000.0; 
     show.dT = currentTime - show.timeStamp;
     show.timeStamp = currentTime;
